feat: move useIntegrationHook file to composables folder
