+++ app/javascript/dashboard/routes/dashboard/settings/integrations/NewHook.vue
@@ -2,7 +2,7 @@
 <script>
 import { mapGetters } from 'vuex';
 import { useAlert } from 'dashboard/composables';
-import { useIntegrationHook } from './useIntegrationHook';
+import { useIntegrationHook } from 'dashboard/composables/useIntegrationHook';
 
 export default {
   props: {
