+++ app/javascript/dashboard/routes/dashboard/settings/integrations/SingleIntegrationHooks.vue
@@ -1,5 +1,5 @@
 <script>
-import { useIntegrationHook } from './useIntegrationHook';
+import { useIntegrationHook } from 'dashboard/composables/useIntegrationHook';
 export default {
   props: {
     integration: {
